changes floats to double and rename to sciscoreinput

+++ src/dotnet/CarbonAware.Aggregators/src/CarbonAware/ISciScoreAggregator.cs
@@ -7,12 +7,12 @@ namespace CarbonAware.Aggregators.SciScore
     public interface ISciScoreAggregator : IAggregator
     {
         /// <summary>
-        /// Returns a float that is made up of 
+        /// Returns a double that is made up of 
         /// </summary>
         /// <param name="location">Location object representing the location for the carbon intensity data.</param>
         /// <param name="timeInterval">ISO8601 time interval representing the interval to calculate carbon intensity over.</param>
         /// <returns>An IEnumerable instance with EmissionsData instances.</returns>
         // Location, 
-        public Task<float> CalculateAverageCarbonIntensityAsync(Location location, string timeInterval);
+        public Task<double> CalculateAverageCarbonIntensityAsync(Location location, string timeInterval);
     }
 }
