+++ src/dotnet/CarbonAware/Model/SciScore.cs
@@ -3,9 +3,9 @@
 [Serializable]
 public record SciScore
 {
-    public float? SciScoreValue { get; set; }
-    public float? EnergyValue { get; set; }
-    public float? MarginalCarbonEmissionsValue { get; set; }
-    public float? EmbodiedEmissionsValue { get; set; }
+    public double? SciScoreValue { get; set; }
+    public double? EnergyValue { get; set; }
+    public double? MarginalCarbonEmissionsValue { get; set; }
+    public double? EmbodiedEmissionsValue { get; set; }
     public Int64? FunctionalUnitValue { get; set; }
 }
